REFS Maintenance - Adds helper for golden sseed.
REFS maintenance - removes something.

Stuff
